+++ src/ids-base/ids-events-mixin.js
@@ -49,7 +49,8 @@ const IdsEventsMixin = (superclass) => class extends IdsRenderLoopMixin(supercla
    * Remove event listener
    * @param {string} eventName The event name with optional namespace
    * @param {HTMLElement} target The DOM element to register
-   * @param {object} options Additional event settings (passive, once, bubbles ect)
+   * @param {object} options Additional event settings (passive, once, passive ect)
+   * @param {HTMLElement} target The DOM element to deregister, defaults to what was already registered
    */
   offEvent(eventName, target, options) {
     const handler = this.handledEvents.get(eventName);
@@ -71,8 +72,9 @@ const IdsEventsMixin = (superclass) => class extends IdsRenderLoopMixin(supercla
       return;
     }
 
-    if (handler?.callback) {
-      target.removeEventListener(eventName.split('.')[0], handler.callback, options || handler.options);
+    const targetApplied = target || handler?.target;
+    if (handler?.callback && targetApplied?.removeEventListener) {
+      targetApplied.removeEventListener(eventName.split('.')[0], handler.callback, options || handler.options);
     }
   }