+++ src/ids-wizard/ids-wizard-step.js
@@ -1,0 +1,20 @@
+import {
+  IdsElement,
+  customElement
+} from '../ids-base/ids-element';
+
+// Note: this component is only used
+// to count steps and retrieve attributes
+// via parent for markup for simplicity sake;
+// does not actually render it's own markup
+
+/**
+ * IDS WizardStep Component
+ *
+ * @type {IdsWizardStep}
+ * @inherits IdsElement
+ */
+@customElement('ids-wizard-step')
+class IdsWizardStep extends IdsElement {}
+
+export default IdsWizardStep;